+++ ipv-api/src/main/java/uk/gov/di/authentication/ipv/lambda/IPVAuthorisationHandler.java
@@ -4,14 +4,11 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
-import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.oauth2.sdk.AuthorizationRequest;
 import com.nimbusds.oauth2.sdk.ParseException;
 import com.nimbusds.oauth2.sdk.ResponseType;
 import com.nimbusds.oauth2.sdk.id.ClientID;
 import com.nimbusds.oauth2.sdk.id.State;
-import com.nimbusds.oauth2.sdk.id.Subject;
-import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.openid.connect.sdk.AuthenticationRequest;
 import com.nimbusds.openid.connect.sdk.OIDCClaimsRequest;
 import com.nimbusds.openid.connect.sdk.claims.ClaimsSetRequest;
@@ -22,7 +19,6 @@ import uk.gov.di.authentication.ipv.entity.IPVAuthorisationRequest;
 import uk.gov.di.authentication.ipv.entity.IPVAuthorisationResponse;
 import uk.gov.di.authentication.ipv.entity.IPVCallbackNoSessionException;
 import uk.gov.di.authentication.ipv.services.IPVAuthorisationService;
-import uk.gov.di.authentication.ipv.services.StorageTokenService;
 import uk.gov.di.orchestration.shared.entity.ClientRegistry;
 import uk.gov.di.orchestration.shared.entity.ErrorResponse;
 import uk.gov.di.orchestration.shared.entity.UserProfile;
@@ -45,7 +41,6 @@ import uk.gov.di.orchestration.shared.services.RedisConnectionService;
 import uk.gov.di.orchestration.shared.services.SessionService;
 import uk.gov.di.orchestration.shared.state.UserContext;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -67,7 +62,6 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
     private final IPVAuthorisationService authorisationService;
     private final NoSessionOrchestrationService noSessionOrchestrationService;
     private final CloudwatchMetricsService cloudwatchMetricsService;
-    private final StorageTokenService storageTokenService;
 
     public IPVAuthorisationHandler(
             ConfigurationService configurationService,
@@ -78,8 +72,7 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
             AuditService auditService,
             IPVAuthorisationService authorisationService,
             NoSessionOrchestrationService noSessionOrchestrationService,
-            CloudwatchMetricsService cloudwatchMetricsService,
-            StorageTokenService storageTokenService) {
+            CloudwatchMetricsService cloudwatchMetricsService) {
         super(
                 IPVAuthorisationRequest.class,
                 configurationService,
@@ -91,7 +84,6 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
         this.authorisationService = authorisationService;
         this.noSessionOrchestrationService = noSessionOrchestrationService;
         this.cloudwatchMetricsService = cloudwatchMetricsService;
-        this.storageTokenService = storageTokenService;
     }
 
     public IPVAuthorisationHandler() {
@@ -109,7 +101,6 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
         this.noSessionOrchestrationService =
                 new NoSessionOrchestrationService(configurationService);
         this.cloudwatchMetricsService = new CloudwatchMetricsService(configurationService);
-        this.storageTokenService = new StorageTokenService(configurationService);
     }
 
     @Override
@@ -143,7 +134,7 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
                             configurationService.getInternalSectorUri(),
                             authenticationService);
             var state = new State();
-            var claimsSetRequest = buildIpvClaimsRequest(authRequest, pairwiseSubject).orElse(null);
+            var claimsSetRequest = buildIpvClaimsRequest(authRequest).orElse(null);
 
             var clientSessionId =
                     getHeaderValueFromHeaders(
@@ -230,25 +221,9 @@ public class IPVAuthorisationHandler extends BaseFrontendHandler<IPVAuthorisatio
         }
     }
 
-    private Optional<ClaimsSetRequest> buildIpvClaimsRequest(
-            AuthenticationRequest authRequest, Subject internalPairwiseSubject) {
-
-        Optional<ClaimsSetRequest> claimsSetRequest =
-                Optional.ofNullable(authRequest)
-                        .map(AuthenticationRequest::getOIDCClaims)
-                        .map(OIDCClaimsRequest::getUserInfoClaimsRequest);
-
-        if (configurationService.isStorageTokenToIpvEnabled()) {
-            AccessToken storageToken =
-                    storageTokenService.generateAndSignStorageToken(
-                            internalPairwiseSubject, JWSAlgorithm.ES256);
-            claimsSetRequest.ifPresent(
-                    csr ->
-                            csr.add(
-                                    new ClaimsSetRequest.Entry(
-                                                    configurationService.getStorageTokenClaimName())
-                                            .withValues(List.of(storageToken.getValue()))));
-        }
-        return claimsSetRequest;
+    private Optional<ClaimsSetRequest> buildIpvClaimsRequest(AuthenticationRequest authRequest) {
+        return Optional.ofNullable(authRequest)
+                .map(AuthenticationRequest::getOIDCClaims)
+                .map(OIDCClaimsRequest::getUserInfoClaimsRequest);
     }
 }
