+++ oidc-api/src/main/java/uk/gov/di/authentication/oidc/services/InitiateIPVAuthorisationService.java
@@ -2,10 +2,13 @@ package uk.gov.di.authentication.oidc.services;
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+import com.nimbusds.jose.JWSAlgorithm;
 import com.nimbusds.oauth2.sdk.AuthorizationRequest;
 import com.nimbusds.oauth2.sdk.ResponseType;
 import com.nimbusds.oauth2.sdk.id.ClientID;
 import com.nimbusds.oauth2.sdk.id.State;
+import com.nimbusds.oauth2.sdk.id.Subject;
+import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.openid.connect.sdk.AuthenticationRequest;
 import com.nimbusds.openid.connect.sdk.OIDCClaimsRequest;
 import com.nimbusds.openid.connect.sdk.claims.ClaimsSetRequest;
@@ -14,6 +17,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import uk.gov.di.authentication.ipv.domain.IPVAuditableEvent;
 import uk.gov.di.authentication.ipv.services.IPVAuthorisationService;
+import uk.gov.di.authentication.ipv.services.StorageTokenService;
 import uk.gov.di.orchestration.shared.entity.ClientRegistry;
 import uk.gov.di.orchestration.shared.entity.ResponseHeaders;
 import uk.gov.di.orchestration.shared.entity.Session;
@@ -41,18 +45,21 @@ public class InitiateIPVAuthorisationService {
     private final IPVAuthorisationService authorisationService;
     private final CloudwatchMetricsService cloudwatchMetricsService;
     private final NoSessionOrchestrationService noSessionOrchestrationService;
+    private final StorageTokenService storageTokenService;
 
     public InitiateIPVAuthorisationService(
             ConfigurationService configurationService,
             AuditService auditService,
             IPVAuthorisationService authorisationService,
             CloudwatchMetricsService cloudwatchMetricsService,
-            NoSessionOrchestrationService noSessionOrchestrationService) {
+            NoSessionOrchestrationService noSessionOrchestrationService,
+            StorageTokenService storageTokenService) {
         this.configurationService = configurationService;
         this.auditService = auditService;
         this.authorisationService = authorisationService;
         this.cloudwatchMetricsService = cloudwatchMetricsService;
         this.noSessionOrchestrationService = noSessionOrchestrationService;
+        this.storageTokenService = storageTokenService;
     }
 
     public APIGatewayProxyResponseEvent sendRequestToIPV(
@@ -76,7 +83,7 @@ public class InitiateIPVAuthorisationService {
         var pairwiseSubject = userInfo.getSubject();
 
         var state = new State();
-        var claimsSetRequest = buildIpvClaimsRequest(authRequest).orElse(null);
+        var claimsSetRequest = buildIpvClaimsRequest(authRequest, pairwiseSubject).orElse(null);
 
         var encryptedJWT =
                 authorisationService.constructRequestJWT(
@@ -127,9 +134,26 @@ public class InitiateIPVAuthorisationService {
                 null);
     }
 
-    private Optional<ClaimsSetRequest> buildIpvClaimsRequest(AuthenticationRequest authRequest) {
-        return Optional.ofNullable(authRequest)
-                .map(AuthenticationRequest::getOIDCClaims)
-                .map(OIDCClaimsRequest::getUserInfoClaimsRequest);
+    private Optional<ClaimsSetRequest> buildIpvClaimsRequest(
+            AuthenticationRequest authRequest, Subject internalPairwiseSubject) {
+
+        Optional<ClaimsSetRequest> claimsSetRequest =
+                Optional.ofNullable(authRequest)
+                        .map(AuthenticationRequest::getOIDCClaims)
+                        .map(OIDCClaimsRequest::getUserInfoClaimsRequest);
+
+        if (configurationService.isStorageTokenToIpvEnabled()) {
+            LOG.info("Adding storageAccessToken claim to IPV claims request");
+            AccessToken storageToken =
+                    storageTokenService.generateAndSignStorageToken(
+                            internalPairwiseSubject, JWSAlgorithm.ES256);
+            claimsSetRequest.ifPresent(
+                    csr ->
+                            csr.add(
+                                    new ClaimsSetRequest.Entry(
+                                                    configurationService.getStorageTokenClaimName())
+                                            .withValues(List.of(storageToken.getValue()))));
+        }
+        return claimsSetRequest;
     }
 }
